clean: unified import_writing_statement
