+++ lib/component/local/attendance_modal/attendance_modal.dart
@@ -1,19 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../domain/class/class.dart';
-import '../../../domain/student/student.dart';
-import '../../../domain/value/class_name.dart';
-import '../../../page/lesson_detail_page.dart';
-import '../../shered/navigator.dart';
-import '../../shered/riverpod.dart';
-import '../../shered/single/buttons.dart';
-import '../../shered/single/color.dart';
-import '../../shered/single/custom_icon_button.dart';
-import '../../shered/single/department_tag.dart';
-import '../../shered/single/space_box.dart';
-import '../../shered/single/test_style.dart';
+
+import '../../shared/navigator.dart';
+import '../../shared/riverpod.dart';
+import '../../shared/single/buttons.dart';
+import '../../shared/single/color.dart';
+import '../../shared/single/custom_icon_button.dart';
+import '../../shared/single/department_tag.dart';
+import '../../shared/single/space_box.dart';
+import '../../shared/single/test_style.dart';
 import 'hooks.dart';
+import '/domain/class/class.dart';
+import '/domain/student/student.dart';
+import '/domain/value/class_name.dart';
+import '/page/lesson_detail_page.dart';
 
 Future attendanceModal(BuildContext context, Class classInfo, WidgetRef ref,
     Student student, ClassName className) {
