+++ lib/component/local/timetable_content/element/element/timetable_header_row.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../shered/single/color.dart';
-import '../../../../shered/single/test_style.dart';
+
 import 'package:intl/intl.dart';
 import 'package:intl/date_symbol_data_local.dart';
+import '../../../../shared/single/color.dart';
+import '../../../../shared/single/test_style.dart';
 
 class TimetableHeaderRow extends HookWidget {
   const TimetableHeaderRow({Key? key, required this.weakDay}) : super(key: key);
