+++ lib/component/local/timetable_content/element/timetable_content_case_of_basic_c.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../domain/class/class.dart';
-import '../../../shered/single/space_box.dart';
-import '../../lesson_detail_body/hooks.dart';
+
+import '../../../shared/single/space_box.dart';
+import '/component/local/lesson_detail_body/hooks.dart';
+import '/domain/class/class.dart';
 import 'element/lesson.dart';
 import 'element/timetable_header_row.dart';
