+++ lib/component/local/timetable_content/timetable_content.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../shered/single/space_box.dart';
+
+import '../../shared/single/space_box.dart';
 import 'element/element/timetable_header_column.dart';
 import 'element/timetable_content_case_of_applied_class.dart';
 import 'hook.dart';
