+++ lib/component/shared/riverpod.dart
@@ -1,10 +1,11 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:vantan_connect/domain/class_document/class_document.dart';
-import 'package:vantan_connect/domain/student/student.dart';
-import 'package:vantan_connect/infrastructure/student_repository.dart';
-import '../../infrastructure/query_service_repository.dart';
-import '../../use_case/actor/student.dart';
-import '../../use_case/query_survice/query_service.dart';
+
+import '/domain/class_document/class_document.dart';
+import '/domain/student/student.dart';
+import '/infrastructure/student_repository.dart';
+import '/infrastructure/query_service_repository.dart';
+import '/use_case/actor/student.dart';
+import '/use_case/query_survice/query_service.dart';
 
 final studentUseCase = Provider((ref) => StudentUseCase(StudentRepository()));
