+++ lib/component/shared/single/buttons.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:vantan_connect/component/shered/single/space_box.dart';
-import 'color.dart';
+
+import '../../shared/single/color.dart';
+import '../../shared/single/space_box.dart';
 
 class FilledEnabledButton extends StatelessWidget {
   const FilledEnabledButton({
