+++ lib/component/shared/single/chips.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:vantan_connect/component/shered/single/space_box.dart';
-import 'package:vantan_connect/component/shered/single/test_style.dart';
+
+import '../../shared/single/space_box.dart';
+import '../../shared/single/test_style.dart';
 
 class SelectedChips extends StatelessWidget {
   const SelectedChips({
