+++ lib/component/shared/single/original_toggle_buttons.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:vantan_connect/component/shered/single/space_box.dart';
+
+import '../../shared/single/space_box.dart';
 
 class OriginalToggleIconText extends StatelessWidget {
   const OriginalToggleIconText({
