+++ lib/component/shared/single/registerd_lesson_by_student_mark.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'color.dart';
-import 'test_style.dart';
+
+import '../../shared/single/color.dart';
+import '../../shared/single/test_style.dart';
 
 class RegisteredLessonByStudentMark extends StatelessWidget {
   const RegisteredLessonByStudentMark({Key? key}) : super(key: key);
