+++ lib/infrastructure/class_repository.dart
@@ -1,10 +1,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:vantan_connect/domain/class/class.dart';
-import 'package:vantan_connect/domain/class_document/class_document.dart';
-import 'package:vantan_connect/domain/student/student.dart';
-import 'package:vantan_connect/domain/teacher/teacher.dart';
-import 'package:vantan_connect/domain/value/student_id.dart';
-import 'package:vantan_connect/domain/value/teacher_id.dart';
+
+import '/domain/class/class.dart';
+import '/domain/class_document/class_document.dart';
+import '/domain/student/student.dart';
+import '/domain/teacher/teacher.dart';
+import '/domain/value/student_id.dart';
+import '/domain/value/teacher_id.dart';
 
 class ClassRepository extends IClassRepository {
   final firestore = FirebaseFirestore.instance;
