+++ lib/infrastructure/executive_stuff_repository.dart
@@ -1,7 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:vantan_connect/domain/executive_stuff/executive_stuff.dart';
-import 'package:vantan_connect/domain/school_term/school_term.dart';
-import 'package:vantan_connect/use_case/actor/executive_stuff.dart';
+
+import '/domain/executive_stuff/executive_stuff.dart';
+import '/domain/school_term/school_term.dart';
+import '/use_case/actor/executive_stuff.dart';
 
 class ExecutiveStuffRepository extends IExecutiveStuffRepository {
   final firestore = FirebaseFirestore.instance;
