+++ lib/infrastructure/query_service_repository.dart
@@ -1,7 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:vantan_connect/domain/class_document/class_document.dart';
-import 'package:vantan_connect/domain/student/student.dart';
-import 'package:vantan_connect/query_service/query_service.dart';
+
+import '/domain/class_document/class_document.dart';
+import '/domain/student/student.dart';
+import '/query_service/query_service.dart';
 
 class QueryServiceRepositoryWhichUseFirebase extends IQueryService {
   final firestore = FirebaseFirestore.instance;
