+++ lib/main.dart
@@ -2,7 +2,8 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:vantan_connect/page/test_app.dart';
+
+import '/page/test_app.dart';
 import 'firebase_options.dart';
 
 Future<void> main() async {
