+++ lib/page/attendance_book_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../component/shered/riverpod.dart';
-import '../domain/class_document/class_document.dart';
+
+import '../component/shared/riverpod.dart';
+import '/domain/class_document/class_document.dart';
 
 class AttendanceBookPage extends ConsumerWidget {
   const AttendanceBookPage({Key? key, required this.classDocument})
