+++ lib/page/lesson_detail_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import '../component/local/lesson_detail_body/lesson_detail_body.dart';
-import '../component/local/lesson_detail_header/lesson_detail_header.dart';
-import '../domain/class/class.dart';
+
+import '/component/local/lesson_detail_body/lesson_detail_body.dart';
+import '/component/local/lesson_detail_header/lesson_detail_header.dart';
+import '/domain/class/class.dart';
 
 class LessonDetailPage extends StatelessWidget {
   const LessonDetailPage({Key? key, required this.classInfo}) : super(key: key);
