+++ lib/page/test_app.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:vantan_connect/page/timetable.dart';
-import '../component/shered/single/color.dart';
+
+import '../component/shared/single/color.dart';
+import '/page/timetable.dart';
 
 class TestApp extends HookWidget {
   TestApp({Key? key}) : super(key: key);
