+++ lib/page/timetable.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../component/local/timetable_content/timetable_content.dart';
-import '../component/local/timetable_header/timetable_header.dart';
-import '../component/shered/single/space_box.dart';
+
+import '../component/shared/single/space_box.dart';
+import '/component/local/timetable_content/timetable_content.dart';
+import '/component/local/timetable_header/timetable_header.dart';
 
 class TimeTablePage extends StatelessWidget {
   const TimeTablePage({Key? key}) : super(key: key);
