+++ lib/use_case/actor/executive_stuff.dart
@@ -1,5 +1,5 @@
-import 'package:vantan_connect/domain/executive_stuff/executive_stuff.dart';
-import 'package:vantan_connect/domain/school_term/school_term.dart';
+import '/domain/executive_stuff/executive_stuff.dart';
+import '/domain/school_term/school_term.dart';
 
 class ExecutiveStaffUseCase {
   ExecutiveStaffUseCase(this.executiveStuffRepository);
