+++ lib/use_case/actor/student.dart
@@ -1,6 +1,6 @@
-import 'package:vantan_connect/component/local/attendance_modal_with_reason/elements/radio_button_with_attendance_status/hooks/useAttendanceState.dart';
-import 'package:vantan_connect/domain/student/student.dart';
-import 'package:vantan_connect/domain/value/class_name.dart';
+import '/component/local/attendance_modal_with_reason/elements/radio_button_with_attendance_status/hooks/useAttendanceState.dart';
+import '/domain/student/student.dart';
+import '/domain/value/class_name.dart';
 
 class StudentUseCase {
   StudentUseCase(this.studentRepository);
