+++ lib/use_case/query_survice/query_service.dart
@@ -1,6 +1,6 @@
-import 'package:vantan_connect/domain/class_document/class_document.dart';
-import 'package:vantan_connect/domain/student/student.dart';
-import '../../query_service/query_service.dart';
+import '/domain/class_document/class_document.dart';
+import '/domain/student/student.dart';
+import '/query_service/query_service.dart';
 
 class QueryServiceUseCase {
   QueryServiceUseCase({required this.queryService});
