+++ lib/use_case/state/base_class_notifier.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../../domain/class/class.dart';
-import '../../infrastructure/class_repository.dart';
+
+import '/domain/class/class.dart';
+import '/infrastructure/class_repository.dart';
 
 class BaseClassNotifier extends StateNotifier<List<Class>> {
   BaseClassNotifier(this.classRepository) : super([]);
