+++ lib/use_case/state/class_document.dart
@@ -1,9 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:vantan_connect/domain/class_document/class_document.dart';
-import 'package:vantan_connect/infrastructure/fake_class_repository.dart';
-import 'package:vantan_connect/use_case/state/class_use_case.dart';
-import '../../domain/class/class.dart';
-import '../../infrastructure/class_repository.dart';
+
+import '/domain/class/class.dart';
+import '/domain/class_document/class_document.dart';
+import '/infrastructure/fake_class_repository.dart';
+import '/use_case/state/class_use_case.dart';
+import '/infrastructure/class_repository.dart';
 
 class ClassDocumentNotifier extends StateNotifier<List<ClassDocument>> {
   ClassDocumentNotifier(this.classRepository) : super([]);
