+++ lib/use_case/state/class_use_case.dart
@@ -1,11 +1,12 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:vantan_connect/domain/class_document/class_document.dart';
-import 'package:vantan_connect/domain/teacher/teacher.dart';
-import 'package:vantan_connect/domain/value/student_id.dart';
-import 'package:vantan_connect/domain/value/teacher_id.dart';
-import '../../domain/class/class.dart';
-import '../../domain/student/student.dart';
-import '../../infrastructure/class_repository.dart';
+
+import '/domain/class_document/class_document.dart';
+import '/domain/teacher/teacher.dart';
+import '/domain/value/student_id.dart';
+import '/domain/value/teacher_id.dart';
+import '/domain/class/class.dart';
+import '/domain/student/student.dart';
+import '/infrastructure/class_repository.dart';
 
 final testTeacherList = [
   Teacher(
