+++ lib/use_case/state/selectable_class_notifier.dart
@@ -1,7 +1,8 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:vantan_connect/use_case/state/class_use_case.dart';
-import '../../domain/class/class.dart';
-import '../../infrastructure/class_repository.dart';
+
+import '/use_case/state/class_use_case.dart';
+import '/domain/class/class.dart';
+import '/infrastructure/class_repository.dart';
 
 class SelectableClassNotifier extends StateNotifier<List<Class>> {
   SelectableClassNotifier(this.classRepository) : super([]);
